requirements.txt
Django==4.1.5
phonenumbers==8.13.4
python-dotenv==0.21.0
python-telegram-bot==13.14

